requirements.txt
base58==0.2.2
boto3==1.2.5
botocore==1.3.30
docutils==0.12
Flask==0.10.1
futures==3.0.4
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
lambda-packages==0.4.0
MarkupSafe==0.23
Pillow==3.1.1
python-dateutil==2.4.2
PyYAML==3.11
requests==2.9.1
six==1.10.0
tqdm==3.7.1
Werkzeug==0.11.4
wheel==0.24.0
wsgi-request-logger==0.4.4
